Updare start script
The developer dashboard takes a long time to load since it is loaded in the browser, and I cannot seem to download the app unless the connection is set to tunnel. By setting this flag, development will be faster for those who want disable the developer dashboard and work straight from the terminal.